+++ src/components/fonts.css
@@ -1,0 +1,1 @@
+@import url('https://fonts.googleapis.com/css?family=Montserrat:400,700');
