+++ src/pages/404.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import Layout from '../components/layout'
+import Theme from '../components/theme'
 
 const NotFoundPage = () => (
-  <Layout>
+  <Theme>
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </Layout>
+  </Theme>
 )
 
 export default NotFoundPage
