+++ src/pages/page-2.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
-import Layout from '../components/layout'
+import Theme from '../components/theme'
 
 const SecondPage = () => (
-  <Layout>
+  <Theme>
     <h1>Hi from the second page</h1>
     <p>Welcome to page 2</p>
     <Link to="/">Go back to the homepage</Link>
-  </Layout>
+  </Theme>
 )
 
 export default SecondPage
